requirements.txt
flask==2.0.3
numpy==1.23.5
opencv-python==4.5.5.64
Pillow==9.1.1